backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
PyJWT==2.8.0
influxdb-client==1.36.0
requests==2.31.0
python-multipart==0.0.6